+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/client/LevelRendererMixin.java
@@ -1,5 +1,6 @@
 package insane96mcp.iguanatweaksreborn.mixin.client;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import insane96mcp.iguanatweaksreborn.module.client.Misc;
@@ -11,9 +12,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(LevelRenderer.class)
 public abstract class LevelRendererMixin {
@@ -25,19 +24,13 @@ public abstract class LevelRendererMixin {
 		return value;
 	}
 
-	@ModifyVariable(at = @At(value = "STORE", ordinal = 2), method = "renderWorldBorder", ordinal = 1)
-	private double onWorldBorderAlpha(double value) {
-		return value * WorldBorder.getTransparencyMultiplier();
-	}
-
-	@Inject(method = "renderHitOutline", at = @At(value = "HEAD"), cancellable = true)
-	private void onRenderHitOutline(PoseStack pPoseStack, VertexConsumer pConsumer, Entity entity, double pCamX, double pCamY, double pCamZ, BlockPos pPos, BlockState state, CallbackInfo ci) {
+	@ModifyExpressionValue(method = "renderHitOutline", at = @At(value = "CONSTANT", args = "floatValue=0.0", ordinal = 0))
+	private float onHitOutline(float value, PoseStack pPoseStack, VertexConsumer pConsumer, Entity entity, double pCamX, double pCamY, double pCamZ, BlockPos pPos, BlockState state) {
 		if (!(entity instanceof Player player)
 				|| player.getAbilities().instabuild
 				|| !state.requiresCorrectToolForDrops()
-				|| player.hasCorrectToolForDrops(state)
-				|| !Misc.shouldHideBlockBreakOutline())
-			return;
-		ci.cancel();
+				|| player.hasCorrectToolForDrops(state))
+			return value;
+		return Misc.getRedOutlineAmount(value);
 	}
 }